tests/wpt/css-tests/css-flexbox-1_dev/html/flexbox_computedstyle_flex-flow-column-wrap-reverse.htm

<!DOCTYPE html>
<html><head><title>flexbox | computed style | flex-flow: column wrap-reverse</title>
<link href="http://opera.com" rel="author" title="Opera Software">
<link href="http://www.w3.org/TR/css-flexbox-1/#propdef-flex-flow" rel="help">
<link href="http://www.w3.org/TR/css-flexbox-1/#valdef-flex-direction-column" rel="help">
<link href="http://www.w3.org/TR/css-flexbox-1/#valdef-flex-wrap-wrap-reverse" rel="help">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="dom" name="flags">
<style>
body {
	color: red;
	flex-flow: column wrap-reverse;
}
</style>
</head><body><div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("flex-direction"),
	          "column");
    assert_equals(getComputedStyle(body).getPropertyValue("flex-wrap"),
	          "wrap-reverse");
});
</script>
</body></html>
